% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saccades.R
\name{add_scanpath.data.frame}
\alias{add_scanpath.data.frame}
\title{Add Scanpath to a Data Frame}
\usage{
\method{add_scanpath}{data.frame}(x, outvar = "scanpath", fixvar = "fixgroup", ...)
}
\arguments{
\item{x}{A data frame.}

\item{outvar}{The output variable name for the scanpath. Defaults to "scanpath".}

\item{fixvar}{The fixation group variable name. Defaults to "fixgroup".}

\item{...}{Additional arguments (currently unused).}
}
\value{
A data frame with the added scanpath.
}
\description{
This function adds a scanpath to a data frame.
}
\examples{
fg <- fixation_group(x = 1:5, y = 6:10, duration = rep(0.2, 5), onset = 1:5)
df <- tibble::tibble(fixgroup = list(fg))
df <- add_scanpath(df)
}
