% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{as.data.frame.eye_density}
\alias{as.data.frame.eye_density}
\title{Convert an eye_density object to a data.frame.}
\usage{
\method{as.data.frame}{eye_density}(x, ...)
}
\arguments{
\item{x}{An eye_density object to be converted into a data.frame.}

\item{...}{Additional arguments passed to the method (currently not used).}
}
\value{
A data.frame with columns x, y, and z representing the x-axis, y-axis, and density values, respectively.
}
\description{
This function converts an eye_density object into a data.frame with x, y, and z values.
}
\details{
The function extracts the x and y values from the eye_density object, then creates a data.frame with all possible combinations of x and y using purrr::cross_df(). It then adds a new column 'z' to the data.frame with the density values from the eye_density object.
}
