% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_frame.R
\name{as_eye_table}
\alias{as_eye_table}
\title{Reapply the 'eye_table' Class to an Object}
\usage{
as_eye_table(x)
}
\arguments{
\item{x}{The input object to which the 'eye_table' class should be (re)applied.}
}
\value{
The input object with the 'eye_table' class (re)applied to its class attribute.
}
\description{
The `as_eye_table` function is a simple utility function that reapplies
the 'eye_table' class to a given object if it is not already part of
its class attribute. This function is not intended for serious use.
}
\examples{
# Create a simple data.frame
df <- data.frame(x = 1:5, y = 6:10)

# Apply the 'eye_table' class to df
eye_table_df <- as_eye_table(df)
class(eye_table_df)
}
