% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blurpoints.R
\name{blurpoints}
\alias{blurpoints}
\title{blurpoints}
\usage{
blurpoints(coords, sigma = 50, xbounds, ybounds, resolution = 1,
  weights = rep(1, nrow(coords)), normalize = TRUE)
}
\description{
blurpoints
}
