% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saccades.R
\name{calcangle}
\alias{calcangle}
\title{Calculate the Angle Between Two Vectors}
\usage{
calcangle(x1, x2)
}
\arguments{
\item{x1}{A numeric vector.}

\item{x2}{A numeric vector.}
}
\value{
A numeric value representing the angle between the two vectors in degrees.
}
\description{
This function calculates the angle (in degrees) between two vectors, x1 and x2.
}
\examples{
calcangle(c(1, 2), c(2, 2))
}
