% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saccades.R
\name{cart2pol}
\alias{cart2pol}
\title{Convert Cartesian Coordinates to Polar Coordinates}
\usage{
cart2pol(x, y)
}
\arguments{
\item{x}{A numeric vector representing the x-coordinates.}

\item{y}{A numeric vector representing the y-coordinates.}
}
\value{
A matrix with two columns, where the first column is rho (the radial coordinate)
  and the second column is theta (the angular coordinate).
}
\description{
This function converts Cartesian coordinates (x, y) to polar coordinates (rho, theta).
}
\examples{
cart2pol(c(1, 2), c(2, 2))
}
