% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{center}
\alias{center}
\title{center eye-movements}
\usage{
center(x, origin, ...)
}
\arguments{
\item{x}{the object}

\item{origin}{the origin of the new coordinate system}

\item{...}{extra args}
}
\description{
center eye-movements
}
