% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{density_matrix}
\alias{density_matrix}
\title{density matrix}
\usage{
density_matrix(x, groups, ...)
}
\arguments{
\item{x}{the object}

\item{groups}{grouping variable}

\item{...}{extra args}
}
\description{
density matrix
}
