% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_frame.R
\name{eye_table}
\alias{eye_table}
\title{construct an eye-movement `data.frame`}
\usage{
eye_table(
  x,
  y,
  duration,
  onset,
  groupvar,
  vars = NULL,
  data,
  clip_bounds = c(0, 1280, 0, 1280),
  relative_coords = TRUE
)
}
\arguments{
\item{x}{the x coordinates}

\item{y}{the y coordinates}

\item{duration}{the duration of the fixations}

\item{onset}{the onset of each fixation}

\item{groupvar}{a grouping variable for the fixations}

\item{vars}{additional variables to retain}

\item{data}{the source `data.frame`}

\item{clip_bounds}{the clip bounds for the field of view}

\item{relaive_coords}{compute relative coordinates}
}
\description{
construct an eye-movement `data.frame`
}
