% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaze_weave_spec.R
\name{gaze_warp_none}
\alias{gaze_warp_none}
\title{Declare no geometric registration for GazeWeave}
\usage{
gaze_warp_none()
}
\value{
A `gaze_warp_spec` object.
}
\description{
Declare no geometric registration for GazeWeave
}
