% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{get_density}
\alias{get_density}
\title{get_density}
\usage{
get_density(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{additional arguments}
}
\description{
Get the density matrix from an object.
}
\concept{get_density}
