% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{kde2d_weighted}
\alias{kde2d_weighted}
\title{Weighted 2D KDE with optional weights}
\usage{
kde2d_weighted(x, y, h, n = 25, lims = c(range(x), range(y)), w)
}
\description{
Weighted 2D KDE with optional weights
}
\keyword{internal}
