% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansion.R
\name{match_scale}
\alias{match_scale}
\title{Match Scaling Parameters for Fixation Data}
\usage{
match_scale(
  ref_tab,
  source_tab,
  match_on,
  refvar = "fixgroup",
  sourcevar = "fixgroup",
  window = NULL,
  ...
)
}
\arguments{
\item{ref_tab}{A data frame containing the reference fixation data.}

\item{source_tab}{A data frame containing the source fixation data.}

\item{match_on}{A string specifying the variable name in both `ref_tab` and `source_tab` to match on.}

\item{refvar}{A string specifying the variable name in `ref_tab` containing the reference fixations (default is "fixgroup").}

\item{sourcevar}{A string specifying the variable name in `source_tab` containing the source fixations (default is "fixgroup").}

\item{window}{A numeric vector of length 2 specifying the time window to restrict the fixations in the source fixation data (default is NULL, which considers all fixations).}

\item{...}{Additional arguments passed to the `estimate_scale` function.}
}
\value{
A data frame containing the original source fixation data with additional columns for the matched scaling parameters.
}
\description{
This function matches the scaling parameters of fixation data between a reference table and a source table based on a common matching variable.
}
\examples{
\dontrun{
# Example usage of the match_scale function
# ref_tab and source_tab should be data frames with fixation groups
matched_data <- match_scale(ref_tab, source_tab, match_on = "subject_id")
}

}
