% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{rescale}
\alias{rescale}
\title{rescale spatial coordinats}
\usage{
rescale(x, sx, sy)
}
\arguments{
\item{x}{the object to rescale}

\item{sx}{the x scale factor}

\item{sy}{the y scale factor}
}
\description{
rescale spatial coordinats
}
