% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{run_similarity_analysis}
\alias{run_similarity_analysis}
\title{Run similarity analysis for fixation data}
\usage{
run_similarity_analysis(
  ref_tab,
  source_tab,
  match_on,
  permutations,
  permute_on = NULL,
  method,
  refvar,
  sourcevar,
  window = NULL,
  multiscale_aggregation = "mean",
  ...
)
}
\arguments{
\item{ref_tab}{A data frame containing the reference fixation groups.}

\item{source_tab}{A data frame containing the source fixation groups to be compared with the reference fixation groups.}

\item{match_on}{A column name in both \code{ref_tab} and \code{source_tab} used for matching the fixation groups.}

\item{permutations}{The number of permutations to perform for assessing the significance of similarity values (default: 0, no permutation tests).}

\item{permute_on}{An optional column name for limiting the matching indices in permutation tests (default: NULL).}

\item{method}{The similarity metric to use for comparing fixation groups (e.g., "sinkhorn", "overlap").}

\item{refvar}{A column name in \code{ref_tab} containing the reference fixation groups.}

\item{sourcevar}{A column name in \code{source_tab} containing the source fixation groups.}

\item{window}{An optional numeric vector specifying the temporal window for computing similarity (default: NULL).}

\item{...}{Extra arguments passed to the similarity function.}
}
\description{
This function compares the similarity between each fixation group in a \code{source_tab}
with a matching fixation group in \code{ref_tab} using the specified similarity metric.
Optionally, permutation tests can be performed for assessing the significance of similarity values.
}
\keyword{internal}
