% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{sample_density.density}
\alias{sample_density.density}
\title{Sample a smooth fixation density map with a set of discrete fixations.}
\usage{
\method{sample_density}{density}(x, fix, times = NULL)
}
\arguments{
\item{x}{An object of class "density" representing the smooth fixation density map.}

\item{fix}{A data frame or tibble containing discrete fixations with columns "x", "y", and "onset".}

\item{times}{A vector of numeric values representing the time points at which the density map should be sampled (default is NULL).}
}
\value{
A data frame with columns "z" and "time", where "z" contains the sampled density values and "time" contains the corresponding time points.
}
\description{
This function samples a smooth fixation density map represented by the object \code{x} with a set of discrete fixations provided in \code{fix}.
}
\details{
The function first checks if the \code{times} parameter is NULL. If so, it directly samples the density map using the coordinates of the fixations in the \code{fix} argument. If the \code{times} parameter is provided, the function first calls the \code{sample_fixations} function to generate a new fixation sequence with the specified time points, and then samples the density map using the coordinates of the new fixation sequence. The result is a data frame containing the sampled density values and the corresponding time points.
}
