% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{scanpath}
\alias{scanpath}
\title{Construct a Scanpath of a Fixation Group of Related Objects}
\usage{
scanpath(x, ...)
}
\arguments{
\item{x}{The fixations.}

\item{...}{Extra arguments.}
}
\value{
A scanpath object.
}
\description{
This function creates a scanpath containing polar coordinates (rho, theta) along with
absolute x and y spatial coordinates for a given fixation group.
}
\examples{
# Create a fixation group
fg <- fixation_group(x=c(.1,.5,1), y=c(1,.5,1), onset=1:3, duration=rep(1,3))
# Create a scanpath for the fixation group
sp <- scanpath(fg)
}
