% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{scanpath_similarity}
\alias{scanpath_similarity}
\title{Scanpath Similarity}
\usage{
scanpath_similarity(
  ref_tab,
  source_tab,
  match_on,
  permutations = 0,
  permute_on = NULL,
  method = c("multimatch"),
  refvar = "scanpath",
  sourcevar = "scanpath",
  window = NULL,
  ...
)
}
\arguments{
\item{ref_tab}{A data frame or tibble containing reference density maps.}

\item{source_tab}{A data frame or tibble containing source density maps.}

\item{match_on}{A character string representing the variable used to match density maps between \code{ref_tab} and \code{source_tab}.}

\item{permutations}{A numeric value specifying the number of permutations for the baseline map (default is 10).}

\item{permute_on}{A character string representing the variable used to stratify permutations (default is NULL).}

\item{method}{The similarity method to use. Currently only "multimatch" is supported.}

\item{refvar}{The name of the column containing scanpaths in the reference table (default is "scanpath").}

\item{sourcevar}{The name of the column containing scanpaths in the source table (default is "scanpath").}

\item{window}{An optional temporal window for restricting the scanpath comparison.}

\item{...}{Extra arguments to pass to the `similarity` function.}
}
\value{
A table containing the computed similarities between scanpaths.
}
\description{
Compute similarity between scanpaths in a source table and matching scanpaths in a reference table.
}
