% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_frame.R
\name{[.eye_table}
\alias{[.eye_table}
\title{Subset an 'eye_table' Object}
\usage{
\method{[}{eye_table}(x, i, j, drop = FALSE)
}
\arguments{
\item{x}{The 'eye_table' object to be subsetted.}

\item{i}{Row indices for subsetting.}

\item{j}{Column indices for subsetting.}

\item{drop}{A logical value indicating whether to drop dimensions that have
only one level after subsetting. Defaults to `FALSE`.}
}
\value{
An 'eye_table' object after subsetting.
}
\description{
The `[.eye_table` function is an S3 method for subsetting 'eye_table' objects.
It uses `NextMethod()` to perform the subsetting operation and then reapplies
the 'eye_table' class to the result using the `as_eye_table` function. This
ensures that the returned object still has the 'eye_table' class after subsetting.
}
\examples{
# Create an 'eye_table' object
df <- data.frame(x = 1:5, y = 6:10)
eye_table_df <- as_eye_table(df)

# Subset the 'eye_table' object
subset_eye_table <- eye_table_df[1:3,]
class(subset_eye_table)
}
