% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{template_multireg}
\alias{template_multireg}
\title{Template Multiple Regression}
\usage{
template_multireg(
  source_tab,
  response,
  covars,
  method = c("lm", "rlm", "nnls", "logistic"),
  intercept = TRUE
)
}
\arguments{
\item{source_tab}{A data frame containing the source maps.}

\item{response}{A character string specifying the column name of the response variable.}

\item{covars}{A character vector specifying the column names of the covariates.}

\item{method}{A character vector of available regression methods. Default is c("lm", "rlm", "nnls", "logistic").
The selected method will be used for the regression analysis.
- "lm": Linear regression (default).
- "rlm": Robust linear regression.
- "nnls": Non-negative least squares.
- "logistic": Logistic regression.}

\item{intercept}{A logical value indicating whether to include an intercept term in the model.
Default is TRUE.}
}
\value{
A data frame with the source table augmented with a new column, multireg, containing
  the results of the multiple regression analysis.
}
\description{
This function performs a multiple regression on the provided source table with the specified
response and covariates using the chosen regression method.
}
