% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{template_sample}
\alias{template_sample}
\title{Sample density maps with coordinates derived from fixation groups.}
\usage{
template_sample(
  source_tab,
  template,
  fixgroup = "fixgroup",
  time = NULL,
  outcol = "sample_out"
)
}
\arguments{
\item{source_tab}{the name of the table containing the density map and fixations}

\item{template}{the name of the template density variable}

\item{fixgroup}{the name of the fixation group supplying the spatiotemporal coordinates used to sample the template}

\item{time}{the time points used to extract coordinates from the `fixation_group`}

\item{outcol}{the name of the output variable}
}
\value{
\code{source_tab} with a list column \code{outcol} of sampled values.
  Rows whose template or fixation group is \code{NULL} cannot be sampled and are
  removed, with a warning.
}
\description{
this function extracts the density for any arbitrary time point in a trial. It simply extracts the value of the density map
for the fixation at time t. The fixations are taken from the `fixgroup` variable and may be associated, for example,
with an independent set of trials.
}
