% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wynn_study_image}
\alias{wynn_study_image}
\title{Study image from Wynn et al.}
\format{
An image object (e.g., a matrix or array representing pixel values).
}
\source{
Wynn et al. eye-tracking study.
}
\usage{
wynn_study_image
}
\description{
An image used in the study phase of the Wynn et al. experiment.
}
\examples{
data(wynn_study_image)

}
\keyword{datasets}
