% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saccades.R
\name{add_scanpath.data.frame}
\alias{add_scanpath.data.frame}
\title{Add Scanpath to a Data Frame}
\usage{
\method{add_scanpath}{data.frame}(x, outvar = "scanpath", fixvar = "fixgroup")
}
\arguments{
\item{x}{A data frame.}

\item{outvar}{The output variable name for the scanpath. Defaults to "scanpath".}

\item{fixvar}{The fixation group variable name. Defaults to "fixgroup".}
}
\value{
A data frame with the added scanpath.
}
\description{
This function adds a scanpath to a data frame.
}
\examples{
# Create a data frame with a fixation group
df <- data.frame(x = 1:5, y = 6:10, fixgroup = rep(1, 5))
# Add a scanpath to the data frame
df <- add_scanpath.data.frame(df)
}
