% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saccades.R
\name{add_scanpath.eye_table}
\alias{add_scanpath.eye_table}
\title{Add Scanpath to an Eye Table}
\usage{
\method{add_scanpath}{eye_table}(x, outvar = "scanpath", fixvar = "fixgroup", ...)
}
\arguments{
\item{x}{An eye table object.}

\item{outvar}{The output variable name for the scanpath. Defaults to "scanpath".}

\item{fixvar}{The fixation group variable name. Defaults to "fixgroup".}

\item{...}{Additional arguments (currently unused).}
}
\value{
An eye table object with the added scanpath.
}
\description{
This function adds a scanpath to an eye table.
}
\examples{
# Create an eye table with a fixation group
df <- data.frame(x = 1:5, y = 6:10, fixgroup = rep(1, 5))
eye_table_df <- as_eye_table(df)
# Add a scanpath to the eye table
eye_table_df <- add_scanpath.eye_table(eye_table_df)
}
