% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{add_scanpath}
\alias{add_scanpath}
\title{add scanpath to dataset}
\usage{
add_scanpath(x, ...)
}
\arguments{
\item{x}{the inpuy dataset}

\item{...}{extra args}
}
\description{
add scanpath to dataset
}
