% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{anim_scanpath}
\alias{anim_scanpath}
\title{animate a fixation scanpath with `gganimate`}
\usage{
anim_scanpath(
  x,
  bg_image = NULL,
  xlim = range(x$x),
  ylim = range(x$y),
  alpha = 1,
  anim_over = c("index", "onset"),
  type = c("points", "raster"),
  time_bin = 1
)
}
\arguments{
\item{x}{a `fixation_group` object}

\item{bg_image}{an image file name}
}
\description{
animate a fixation scanpath with `gganimate`
}
