% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/fixations.R
\name{center}
\alias{center}
\alias{center.fixation_group}
\title{Center Eye-Movements in a New Coordinate System}
\usage{
center(x, origin, ...)

\method{center}{fixation_group}(x, origin = NULL, ...)
}
\arguments{
\item{x}{The input object containing the eye-movements to be centered.}

\item{origin}{A vector containing the x and y coordinates of the new coordinate system's origin.}

\item{...}{Additional arguments to be passed to the centering method.}
}
\value{
An object containing the centered eye-movements in the new coordinate system.
}
\description{
This function centers the eye-movements in a new coordinate system with the specified origin.
}
\examples{
\dontrun{
# Example usage of the center function
new_origin <- c(500, 500) # New coordinate system origin
centered_object <- center(input_object, origin = new_origin)
}

}
