% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansion.R
\name{estimate_scale}
\alias{estimate_scale}
\title{Estimate Scaling Parameters for Fixation Data}
\usage{
estimate_scale(x, y, lower = c(0.1, 0.1), upper = c(10, 10), window)
}
\arguments{
\item{x}{A data frame or matrix containing the x coordinates of the first set of fixations.}

\item{y}{A data frame or matrix containing the y coordinates of the second set of fixations.}

\item{lower}{A numeric vector of length 2 specifying the lower bounds for the scaling parameters (default is c(0.1, 0.1)).}

\item{upper}{A numeric vector of length 2 specifying the upper bounds for the scaling parameters (default is c(10, 10)).}

\item{window}{A numeric vector of length 2 specifying the time window to restrict the fixations in `y` (default is NULL, which considers all fixations).}
}
\value{
A list containing the estimated scaling parameters.
}
\description{
This function estimates the scaling parameters for two sets of fixation data using the Hausdorff distance as an optimization objective.
}
\keyword{internal}
