% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{eye_density.fixation_group}
\alias{eye_density.fixation_group}
\title{Compute a density map for a fixation group.}
\usage{
\method{eye_density}{fixation_group}(
  x,
  sigma = 50,
  xbounds = c(min(x$x), max(x$x)),
  ybounds = c(min(x$y), max(x$y)),
  outdim = c(xbounds[2] - xbounds[1], ybounds[2] - ybounds[1]),
  normalize = TRUE,
  duration_weighted = FALSE,
  window = NULL,
  origin = c(0, 0)
)
}
\arguments{
\item{x}{A fixation_group object.}

\item{sigma}{The standard deviation of the kernel. Default is 50.}

\item{xbounds}{The x-axis bounds. Default is the range of x values in the fixation group.}

\item{ybounds}{The y-axis bounds. Default is the range of y values in the fixation group.}

\item{outdim}{The output dimensions of the density map. Default is the difference between the xbounds and ybounds.}

\item{normalize}{Whether to normalize the output map. Default is TRUE.}

\item{duration_weighted}{Whether to weight the fixations by their duration. Default is FALSE.}

\item{window}{The temporal window over which to compute the density map. Default is NULL.}

\item{origin}{The origin of the coordinate system. Default is c(0,0).}
}
\value{
An object of class c("eye_density", "density", "list") containing the computed density map and other relevant information.
}
\description{
This function computes a density map for a given fixation group using kernel density estimation.
}
\details{
The function computes a density map for a given fixation group using the kde2d function from the MASS package. The density map is computed based on the x and y coordinates of the fixations, with optional weighting by their duration. The resulting density map can be normalized if desired.
}
\seealso{
Other eye_density: 
\code{\link{eye_density}()}
}
\concept{eye_density}
