% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{eye_density}
\alias{eye_density}
\title{eye_density}
\usage{
eye_density(x, sigma, xbounds, ybounds, outdim, weights, normalize, ...)
}
\arguments{
\item{x}{A data frame containing the eye fixations.}

\item{sigma}{Numeric, the standard deviation of the kernel used for density estimation.}

\item{xbounds}{Numeric vector of length 2, defining the minimum and maximum x-axis bounds of the density map.}

\item{ybounds}{Numeric vector of length 2, defining the minimum and maximum y-axis bounds of the density map.}

\item{outdim}{Numeric vector of length 2, specifying the number of rows and columns in the output density map.}

\item{weights}{Optional, a numeric vector of fixation weights to be used in the density estimation.}

\item{normalize}{Logical, whether to normalize the output density map such that its values sum to 1.}

\item{...}{Additional arguments passed on to the method.}
}
\value{
An object of class "eye_density", "density", and "list" containing the computed density map and other relevant information.
}
\description{
Compute a density map for a set of eye fixations.
}
\details{
This function uses kernel density estimation to compute a density map of eye fixations.
It takes various parameters to control the computation and output of the density map.
}
\seealso{
\code{\link{eye_density.fixation_group}} for an example of a specific method implementation for this generic.

Other eye_density: 
\code{\link{eye_density.fixation_group}()}
}
\concept{eye_density}
