% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{fixation_similarity}
\alias{fixation_similarity}
\title{fixation_similarity}
\usage{
fixation_similarity(
  ref_tab,
  source_tab,
  match_on,
  permutations = 0,
  permute_on = NULL,
  method = c("sinkhorn", "overlap", "mm_vector", "mm_direction", "mm_length",
    "mm_position", "mm_duration"),
  refvar = "fixgroup",
  sourcevar = "fixgroup",
  window = NULL,
  ...
)
}
\description{
compute similarity between each fixation group in a \code{source_tab} with a matching fixation group in \code{ref_tab}
}
