% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimatch.R
\name{multi_match}
\alias{multi_match}
\title{multi_match}
\usage{
multi_match(x, y, screensize)
}
\arguments{
\item{x}{the first \code{scanpath}}

\item{y}{the second \code{scanpath}}

\item{screensize}{a two element vector indicating screen size}
}
\description{
multi_match
}
