% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{normalize}
\alias{normalize}
\title{normalize eye-movements to unit range}
\usage{
normalize(x, xbounds, ybounds, ...)
}
\arguments{
\item{x}{the object}

\item{xbounds}{the x bounds}

\item{ybounds}{the y bounds}

\item{...}{extra args}
}
\description{
normalize eye-movements to unit range
}
