% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.eye_density}
\alias{plot.eye_density}
\title{Plot Eye Density}
\usage{
\method{plot}{eye_density}(
  x,
  alpha = 0.8,
  bg_image = NULL,
  transform = c("identity", "sqroot", "curoot", "rank"),
  ...
)
}
\arguments{
\item{x}{An "eye_density" object.}

\item{alpha}{The opacity level for the raster layer (default: 0.8).}

\item{bg_image}{An optional image file name to use as the background.}

\item{transform}{The transformation to apply to the density values (default: c("identity", "sqroot", "curoot", "rank")).}

\item{...}{Additional args}
}
\value{
A ggplot object representing the eye density plot.
}
\description{
This function creates a plot of the eye density using ggplot2.
}
\examples{
# Assume `ed` is an "eye_density" object
# Plot the eye density
plot_eye_density <- plot.eye_density(ed)
}
\seealso{
Other visualization: 
\code{\link{anim_scanpath}()},
\code{\link{plot.fixation_group}()}
}
\concept{visualization}
