% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.fixation_group}
\alias{plot.fixation_group}
\title{plot a fixation_group object}
\usage{
\method{plot}{fixation_group}(
  x,
  type = c("points", "contour", "filled_contour", "density", "raster"),
  bandwidth = 60,
  xlim = range(x$x),
  ylim = range(x$y),
  size_points = TRUE,
  show_points = TRUE,
  show_path = TRUE,
  bins = max(as.integer(length(x$x)/10), 4),
  bg_image = NULL,
  colours = rev(RColorBrewer.brewer.pal(n = 10, "Spectral")),
  alpha_range = c(0.5, 1),
  alpha = 0.8,
  window = NULL,
  transform = c("identity", "sqroot", "curoot", "rank")
)
}
\description{
plot a fixation_group object
}
\examples{

fg <- fixation_group(x=runif(50, 0, 100), y=runif(50, 0, 100), duration=rep(1,50), onset=seq(1,50))
## plot(fg)
}
