% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{rep_fixations}
\alias{rep_fixations}
\title{replicate a fixation sequence}
\usage{
rep_fixations(x, resolution)
}
\arguments{
\item{x}{the object}

\item{resolution}{the temporal resolution of the replicated fixations}
}
\description{
replicate a fixation sequence
}
