% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/expansion.R
\name{rescale}
\alias{rescale}
\alias{rescale.fixation_group}
\title{rescale}
\usage{
rescale(x, sx, sy)

\method{rescale}{fixation_group}(x, sx, sy)
}
\arguments{
\item{x}{An object containing spatial coordinates to be rescaled.}

\item{sx}{A numeric value representing the x-axis scale factor.}

\item{sy}{A numeric value representing the y-axis scale factor.}
}
\value{
An object with rescaled spatial coordinates.
}
\description{
Rescale spatial coordinates.
}
\details{
This function rescales the spatial coordinates of an object by the given scale factors.
}
\seealso{
\code{\link{rescale.fixation_group}} for an example of a specific method implementation for this generic.
}
\concept{rescale}
