% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{sample_density}
\alias{sample_density}
\title{sample a smooth fixation density map with a set of discrete fixations}
\usage{
sample_density(x, fix, ...)
}
\arguments{
\item{x}{the object to sample}

\item{fix}{the fixations}
}
\description{
sample a smooth fixation density map with a set of discrete fixations
}
