% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{sample_fixations}
\alias{sample_fixations}
\title{sample_fixations}
\usage{
sample_fixations(x, time)
}
\arguments{
\item{x}{the fixation group}

\item{time}{the continuous time points to sample}
}
\description{
sample_fixations
}
