% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/fixations.R
\name{sample_fixations}
\alias{sample_fixations}
\alias{sample_fixations.fixation_group}
\title{sample_fixations}
\usage{
sample_fixations(x, time, ...)

\method{sample_fixations}{fixation_group}(x, time, fast = TRUE, ...)
}
\arguments{
\item{x}{the fixation group}

\item{time}{the continuous time points to sample}

\item{...}{Additional arguments passed to methods.}

\item{fast}{Logical. If TRUE (default), uses faster approximation method.}
}
\description{
sample_fixations
}
