% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saccades.R
\name{scanpath.fixation_group}
\alias{scanpath.fixation_group}
\title{Create a Scanpath for a Fixation Group}
\usage{
\method{scanpath}{fixation_group}(x, ...)
}
\arguments{
\item{x}{A fixation group object.}
}
\value{
A scanpath object.
}
\description{
This function creates a scanpath for a fixation group.
}
\examples{
# Create a fixation group
fixgroup <- data.frame(x = 1:5, y = 6:10)
# Create a scanpath for the fixation group
scanpath_obj <- scanpath.fixation_group(fixgroup)
}
