% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{scanpath}
\alias{scanpath}
\title{compute scanpath of a fixation_group of related object}
\usage{
scanpath(x, ...)
}
\arguments{
\item{x}{the fixations}

\item{...}{extra args}
}
\description{
compute scanpath of a fixation_group of related object
}
