% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{similarity}
\alias{similarity}
\title{similarity}
\usage{
similarity(x, y, method, ...)
}
\arguments{
\item{x}{the first object to compare}

\item{y}{the second object to compare}

\item{method}{the similarity metric}
}
\description{
similarity
}
