% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarity.scanpath}
\alias{similarity.scanpath}
\title{Compute Similarity Between Scanpaths}
\usage{
\method{similarity}{scanpath}(
  x,
  y,
  method = c("multimatch"),
  window = NULL,
  screensize = NULL,
  ...
)
}
\arguments{
\item{x}{A data frame containing the first scanpath.}

\item{y}{A data frame containing the second scanpath.}

\item{method}{A character string specifying the method to compute the similarity (default is "multimatch").}

\item{window}{A numeric vector of length 2 specifying the time window to restrict the fixations in the input scanpaths (default is NULL, which considers all fixations).}

\item{screensize}{A numeric vector of length 2 specifying the dimensions of the screen (e.g., c(1000, 1000)). Required for the "multimatch" method.}

\item{...}{Additional arguments passed to the similarity computation method.}
}
\value{
A numeric value representing the similarity between the two input scanpaths.
}
\description{
This function computes the similarity between two scanpaths using a specified method.
}
\examples{
# Example usage of the similarity.scanpath function
scanpath1 <- # first scanpath data
scanpath2 <- # second scanpath data
similarity_value <- similarity.scanpath(scanpath1, scanpath2, method = "multimatch", screensize = c(1000, 1000))

}
\seealso{
Other similarity: 
\code{\link{similarity}()}
}
\concept{similarity}
