% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{template_sample}
\alias{template_sample}
\title{template_sample}
\usage{
template_sample(
  source_tab,
  template,
  fixgroup = "fixgroup",
  time = NULL,
  outcol = "sample_out"
)
}
\description{
template sample extracts the density for any arbitrary time point in a trial. It simply extracts the value of the density map
for the fixation at time t.
}
