% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_frame.R
\name{template_similarity}
\alias{template_similarity}
\title{template_similarity}
\usage{
template_similarity(ref_tab, source_tab, match_on, method = c("spearman",
  "pearson", "cosine"), permutations = 10)
}
\description{
template_similarity
}
