% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{template_similarity}
\alias{template_similarity}
\title{template_similarity}
\usage{
template_similarity(
  ref_tab,
  source_tab,
  match_on,
  permute_on = NULL,
  refvar = "density",
  sourcevar = "density",
  method = c("spearman", "pearson", "fisherz", "cosine", "l1", "jaccard", "dcov", "emd"),
  permutations = 10,
  multiscale_aggregation = "mean",
  ...
)
}
\arguments{
\item{ref_tab}{A data frame or tibble containing reference density maps.}

\item{source_tab}{A data frame or tibble containing source density maps.}

\item{match_on}{A character string representing the variable used to match density maps between \code{ref_tab} and \code{source_tab}.}

\item{permute_on}{A character string representing the variable used to stratify permutations (default is NULL).}

\item{refvar}{A character string representing the name of the variable containing density maps in the reference table (default is "density").}

\item{sourcevar}{A character string representing the name of the variable containing density maps in the source table (default is "density").}

\item{method}{A character string specifying the similarity method to use. Possible values are "spearman", "pearson", "fisherz", "cosine", "l1", "jaccard", and "dcov" (default is "spearman").}

\item{permutations}{A numeric value specifying the number of permutations for the baseline map (default is 10).}

\item{multiscale_aggregation}{If the density maps are multiscale (i.e., `eye_density_multiscale` objects), this specifies how to aggregate similarities from different scales. Options: "mean" (default, returns the average similarity across scales), "none" (returns a list or vector of similarities, one per scale, within the result columns). See `similarity.eye_density_multiscale`.}

\item{...}{Extra arguments to pass to the `similarity` function.}
}
\value{
A data frame or tibble containing the source table and additional columns with the similarity scores and permutation results.
}
\description{
Compute similarity between each density map in a \code{source_tab} with a matching ("template") density map in \code{ref_tab}.
}
