% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{template_similarity}
\alias{template_similarity}
\title{template_similarity}
\usage{
template_similarity(
  ref_tab,
  source_tab,
  match_on,
  refvar = "density",
  sourcevar = "density",
  method = c("spearman", "pearson", "cosine", "l1", "jaccard", "dcov"),
  permutations = 10
)
}
\description{
template_similarity
}
